refactor: split analysis.ts (742 lines) into 5 focused modules
Split domain analysis queries into:
- analysis/domain-analysis.ts (243) — domainAnalysis query
- analysis/value-detail.ts (372) — domainAnalysisValueDetail query
- analysis/value-detail-types.ts (69) — extracted types + helpers
- analysis/condition-transcripts.ts (92) — domainAnalysisConditionTranscripts
- analysis/index.ts (3) — barrel

Codex adversarial review: 5/5 PASS
