+++ cloud/apps/api/src/graphql/queries/domain/index.ts
@@ -2,4 +2,4 @@ import './types.js';
 import './catalog.js';
 import './planning.js';
 import './evaluation/index.js';
-import './analysis.js';
+import './analysis/index.js';
